requirements.txt
fastapi
uvicorn
request
streamlit
pypdf
chromadb
langchain
langchain-core
langchain_community
langchain-openai
openai
python-dotenv



